requirements.txt
fastapi
uvicorn
ffmpeg-python
openai-whisper
transformers
torch